% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AddMsg}
\alias{AddMsg}
\title{Add a Message to a Messages List}
\usage{
AddMsg(messages, type, msg)
}
\arguments{
\item{messages}{a messages list (from NewMessages)}

\item{type}{one of 'output', 'warnings', 'errors'}

\item{msg}{the message string}
}
\value{
the updated messages list
}
\description{
Internal helper: append a message to the appropriate slot
  of a messages list.
}
\keyword{internal}
