% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AdjustEstimatedPlotHeights_obs}
\alias{AdjustEstimatedPlotHeights_obs}
\title{Adjust Estimated Plot Heights}
\usage{
AdjustEstimatedPlotHeights_obs(
  estimated_plot_heights,
  full_height_cm,
  track_height_cm
)
}
\description{
Adjust Estimated Plot Heights
}
\details{
Internal function:
}
