% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{AlignBasicPlotParameters}
\alias{AlignBasicPlotParameters}
\title{Align Basic Plot Parameters}
\usage{
AlignBasicPlotParameters(
  basic_plot_parameters,
  both_strands,
  strands_intermingled,
  fixed_plot_vertical_parameters,
  vertical_parameters,
  minimal_units,
  full_height_cm,
  annotation_packing
)
}
\arguments{
\item{basic_plot_parameters}{}

\item{both_strands}{}

\item{strands_intermingled}{}

\item{fixed_plot_vertical_parameters}{}

\item{vertical_parameters}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
