% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotation_cache_size}
\alias{annotation_cache_size}
\title{Number of entries currently in the annotation-processing cache}
\usage{
annotation_cache_size()
}
\value{
Integer; entry count.
}
\description{
Number of entries currently in the annotation-processing cache
}
\examples{
annotation_cache_size()
}
