% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayRSession.R
\name{as.data.frame.seqNdisplayRSession}
\alias{as.data.frame.seqNdisplayRSession}
\title{Session To Dataframe}
\usage{
\method{as.data.frame}{seqNdisplayRSession}(session, ...)
}
\arguments{
\item{session}{seqNdisplayRSession object}

\item{...}{other arguments to session_to_df}
}
\value{
A data.frame with columns: color, bigwig_directory, bigwig_file, strand, dataset and optionally subgroup_1, subgroup_2, ...
}
\description{
Converts session information to data frame as in Excel import sheet.
}
\details{
see \link{session_to_df} for more details.
}
\examples{

}
