% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{BasicPlotParameters}
\alias{BasicPlotParameters}
\title{Basic Plot Parameters}
\usage{
BasicPlotParameters(
  plotted_strand,
  plotted_region,
  feature_names_font_size,
  plot_height_parameters,
  plot_width_parameters,
  full_width_cm,
  full_height_cm = NULL,
  track_height_cm = 0.3,
  plot_vertical_parameters,
  bin_size,
  bins_per_cm,
  plotting_segment_order,
  tracks_listed,
  unstranded_beds
)
}
\arguments{
\item{plotted_strand}{}

\item{plotted_region}{}

\item{feature_names_font_size}{}

\item{plot_height_parameters}{}

\item{plot_width_parameters}{}

\item{full_width_cm}{}

\item{full_height_cm}{}

\item{track_height_cm}{}

\item{plot_vertical_parameters}{}

\item{bin_size}{}

\item{bins_per_cm}{}

\item{plotting_segment_order}{}

\item{tracks_listed}{}

\item{unstranded_beds}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
