% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ChangeColorLightness}
\alias{ChangeColorLightness}
\title{Change Color Lightness}
\usage{
ChangeColorLightness(hex_color, factor)
}
\arguments{
\item{c}{}
}
\value{
placeholder
}
\description{
Internal function: 
if the lightness value is <0.5 the color is considered dark
if the lightness value is >=0.5 the color is considered light
}
\examples{
c_hex = '#FFA550'
c_new = ChangeColorLightness(c_hex, 0.25)

}
\author{
SLA
}
\keyword{internal}
