% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{ConvertCollapsedFormat}
\alias{ConvertCollapsedFormat}
\title{Convert Collapsed Format}
\usage{
ConvertCollapsedFormat(annot_info)
}
\arguments{
\item{annot_info}{}
}
\value{
placeholder
}
\description{
Internal function: 
This function is the third step in organizing the annotated features in the plotted region into various formats
Used after AnnotatedFeaturesInRegion and OrganizeOverlappingLoci to create a final organized version of the annotated features
Converts the format of "collapsed" and "collapsed2" from GRanges to GRangesLists to make the overall formatting uniform; one GRanges object per 'feature name' (e.g. "TAF1D" and "RP11-178H8.7")
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
