% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{ConvertColor}
\alias{ConvertColor}
\title{Convert Color}
\usage{
ConvertColor(c, phi = 180)
}
\arguments{
\item{c}{}

\item{phi}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
c_hex = '#FFA550'
c_hex_180 = ConvertColor(c_hex)

}
\author{
SLA
}
\keyword{internal}
