% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{CoordsOfFeatName}
\alias{CoordsOfFeatName}
\title{Coords Of Feat Name}
\usage{
CoordsOfFeatName(
  chrom_coord_feature,
  feat_width_cm,
  word_width_cm,
  word_width_feat_fit,
  word_width_plot_fit,
  feature_gr,
  plotted_region
)
}
\description{
Coords Of Feat Name
}
\details{
Internal function: Given a chromosomal coordinate (for the center of the annotated feature name) the parameters necessary for plotting the word such that it fits inside the plotted region are calculated
}
