% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{CoordsOfGenomicScale}
\alias{CoordsOfGenomicScale}
\title{Coords Of Genomic Scale}
\usage{
CoordsOfGenomicScale(
  chrom_coord,
  font_size,
  coords_tracks,
  full_width_cm,
  plot_width,
  plot_start,
  plot_end
)
}
\arguments{
\item{chrom_coord}{}

\item{font_size}{}

\item{coords_tracks}{}

\item{full_width_cm}{}

\item{plot_width}{}

\item{plot_start}{}

\item{plot_end}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
