% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{default_annotation_options}
\alias{default_annotation_options}
\title{Default Annotation Display Options}
\usage{
default_annotation_options()
}
\value{
A named list of default annotation-display options.
}
\description{
Default Annotation Display Options
}
\details{
Default values for options regarding annotation display in seqNdisplayR session object.
}
