% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{default_parameters}
\alias{default_parameters}
\title{Default Dataset-Specific Options}
\usage{
default_parameters()
}
\value{
A named list of default Dataset-specific options.
}
\description{
Default Dataset-Specific Options
}
\details{
Default values for parameters object in seqNdisplayR session object.
}
