% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{deparse_option}
\alias{deparse_option}
\title{Deparse Option}
\usage{
deparse_option(option)
}
\arguments{
\item{option}{named list, vector or single element}
}
\description{
Parse option into string
}
\details{
String representation of object, compatible with parse_option
}
\examples{
deparse_option(c(1.2,3))
deparse_option(list('RNA-seq' = c(1.2,3), 'TT-seq' = c(2,4)))
deparse_option(list('RNA-seq' = c(TRUE,FALSE), 'TT-seq' = c(TRUE,FALSE)))

}
