% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{DeparseOption}
\alias{DeparseOption}
\title{Deparse Option}
\usage{
DeparseOption(option)
}
\arguments{
\item{option}{named list, vector or single element}
}
\value{
String representation of object, compatible with ParseOption
}
\description{
Internal function: 
Parse option into string
}
\examples{
DeparseOption(c(1.2,3))
DeparseOption(list('RNA-seq' = c(1.2,3), 'TT-seq' = c(2,4)))
DeparseOption(list('RNA-seq' = c(TRUE,FALSE), 'TT-seq' = c(TRUE,FALSE)))

}
\author{
MS
}
\keyword{internal}
