% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igv_conversion.R
\name{.clean_prefix}
\alias{.clean_prefix}
\title{Trim trailing whitespace, underscores, dashes from a prefix}
\usage{
.clean_prefix(s)
}
\description{
Trim trailing whitespace, underscores, dashes from a prefix
}
\keyword{internal}
