% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.log_messages}
\alias{.log_messages}
\title{Log messages internally (called by PrintOutput)}
\usage{
.log_messages(messages, caller = NULL)
}
\description{
Log messages internally (called by PrintOutput)
}
\keyword{internal}
