% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Empty2Null}
\alias{Empty2Null}
\title{Empty 2 Null}
\usage{
Empty2Null(x)
}
\arguments{
\item{x}{}
}
\value{
placeholder
}
\description{
Internal function: 
Helper used for parsing parameters$<dataset>$whichSamples
From chatGPT: The purpose of this function is to convert empty elements within a list to NULL. If an element is not a list or has a length of 0, it remains unchanged. However, if an element is a non-empty list, the function recursively processes each nested element to convert empty elements to NULL.
}
\examples{
NULL

}
\author{
MS
}
\keyword{internal}
