% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{empty_df}
\alias{empty_df}
\title{Empty Unnecessary Cells in Sample DataFrame}
\usage{
empty_df(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
Data frame
}
\description{
Empty Unnecessary Cells in Sample DataFrame
}
\details{
Empty rows in data frame from a session data frame such that cells
  which have exactly same entry in cell above are left empty. This only
  applies to columns named 'bigwig_directory', 'dataset' and all columns with
  prefix 'subgroup_'. Output should look similar to the Samples sheet in the
  Excel template in \code{inst/extdata/example_excel_template.xls}.
}
