% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{empty_to_null}
\alias{empty_to_null}
\title{Zero-length object to null}
\usage{
empty_to_null(x)
}
\description{
Zero-length object to null
}
\details{
: Helper used for parsing parameters$<dataset>$whichSamples
}
