% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{EmptyDf}
\alias{EmptyDf}
\title{Empty Df}
\usage{
EmptyDf(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
trimmed dataframe
}
\description{
Internal function: 
Empty Unnecessary Cells in Sample DataFrame
}
\details{
Empty rows in data frame from a session data frame such that cells
  which have exactly same entry in cell above are left empty. This only
  applies to columns named 'bigwig_directory', 'dataset' and all columns with
  prefix 'subgroup_'. Output should look similar to the Samples sheet in the
  Excel template in \code{inst/extdata/example_excel_template.xls}.
}
\examples{
NULL

}
\author{
MS
}
\keyword{internal}
