% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{EvaluateNumericValue}
\alias{EvaluateNumericValue}
\title{Evaluate Numeric Value}
\usage{
EvaluateNumericValue(
  num_val,
  positive_val,
  min_val,
  max_val,
  interval_obligatory = FALSE,
  turn_errors_to_warnings = FALSE,
  alt_par_name = NULL,
  verbosity
)
}
\arguments{
\item{num_val}{}

\item{positive_val}{}

\item{min_val}{}

\item{max_val}{}

\item{interval_obligatory}{}

\item{turn_errors_to_warnings}{}

\item{alt_par_name}{}

\item{verbosity}{}
}
\value{
placeholder
}
\description{
Internal function: 
Evaluate whether a vector/value is numeric and falls within recommmended range
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
