% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ExamplesSampleSheetsFolder}
\alias{ExamplesSampleSheetsFolder}
\title{Examples Sample Sheets Folder}
\usage{
ExamplesSampleSheetsFolder()
}
\value{
finds and returns the folder with example Excel Templates for seqNdisplayR
}
\description{
Finds the folder with example Excel Templates for seqNdisplayR
}
\examples{
example_folder = ExamplesSampleSheetsFolder()
list.files(example_folder)
fname = list.files(example_folder)[7]
xl_fname = paste0(example_folder, fname)
session = LoadExcel(xl_fname, load_annotations = T)
plot(session, feature='LMO4')

}
\author{
SLA
}
