% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{fill_df}
\alias{fill_df}
\title{Fill Empty Rows}
\usage{
fill_df(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
Data frame
}
\description{
Fill Empty Rows
}
\details{
Empty rows in data frame from loaded XL sheet are filled with values
  from columns above in hierarchical fashion such that the right-most columns
  are only filled if specified in the dataset defining row of the seqtype.
  See Excel template sheet in \code{inst/extdata/example_excel_template.xls}
  for more information.
}
