% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{FinalizePlottingSegmentOrder}
\alias{FinalizePlottingSegmentOrder}
\title{Finalize Plotting Segment Order}
\usage{
FinalizePlottingSegmentOrder(
  plotting_segment_order,
  tracks_listed,
  both_strands,
  include_genomic_scale,
  genomic_scale_on_top,
  any_stranded_beds,
  any_unstranded_beds,
  strands_intermingled,
  verbosity,
  interface
)
}
\arguments{
\item{plotting_segment_order}{}

\item{tracks_listed}{}

\item{both_strands}{}

\item{include_genomic_scale}{}

\item{genomic_scale_on_top}{}

\item{any_stranded_beds}{}

\item{any_unstranded_beds}{}

\item{strands_intermingled}{}

\item{verbosity}{}

\item{interface}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
