% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{get_annotations}
\alias{get_annotations}
\title{Annotations from Excel}
\usage{
get_annotations(annotations)
}
\arguments{
\item{annotations}{Data frame}
}
\value{
Named list of path, or named list of GRanges. Names used are entries in the *annotation_name* column.
}
\description{
Reads Annotations from Excel template
}
\details{
annotations is a data.frame with at least 2 columns *annotation_file* and *annotation_name*, both
  case-sensitive and the space is required.
}
