% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{get_colors}
\alias{get_colors}
\title{Get Colors}
\usage{
get_colors(filled_df)
}
\arguments{
\item{filled_df}{a fill data frame (see details)}
}
\value{
Named lists or nested lists of named lists
}
\description{
Get Colors
}
\details{
Get colors from a data frame containing at a minimum columns
  color and dataset. Columns on the right to dataset may contain subcategories,
  called subgroup_1, subgroup_2 etc. UPS: dataset and subgroup_s must be the right-most columns, see example below.
}
\examples{
df <- data.frame(color=c(rep('red', 4), rep('green', 2)),dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), subgroup_2=1:6, stringsAsFactors=FALSE)
get_colors(df)

}
