% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Parameters}
\usage{
get_parameters(samples_df, params_df)
}
\arguments{
\item{samples_df}{a filled data frame with sample information(see details)}

\item{params_df}{dataframe with Dataset_options}
}
\value{
Named list
}
\description{
Get Parameters
}
\details{
Prepare seqNdisplayR parameters based on info in param_df. param_df
  needs to contain a column called dataset, all column names are considered
  names of dataset-specific options. Each row should contain one dataset and
  its information. See examples below. Batch information is from a data frame
  containing the track info. Needs to contain at a minimum columns strand,
  batch and dataset. Strand must be 'plus' or 'minus', use 'plus' for
  unstranded data. If batch is all empty, NAs or all identical, will assume
  no batch correction, otherwise set parameters to batch correction using
  information from the batch column. You can manually change individual
  parameters after this (see bottom examples in the man page).
}
\examples{
samples_df <- data.frame(bigwig_file=c('a.bw', 'b.bw', 'c.bw', 'd.bw', 'e.bw', 'f.bw'),
  strand = rep('plus', 6),
  batch = rep(NA, 6),
  dataset=c(rep('a',4), rep('b',2)),
  subgroup_1=c('x','x','y','y', 'x','y'), stringsAsFactors=FALSE)
params_df <- data.frame(dataset=c('a','b'), calcMean=c('TRUE','FALSE'), log2Transform=c('TRUE','FALSE'), stringsAsFactors=FALSE)
get_parameters(samples_df, params_df)

df <- data.frame(bigwig_file=c('a.bw', 'b.bw', 'c.bw', 'd.bw', 'e.bw', 'f.bw'),
  strand = rep('plus', 6),
  batch = rep(c('rep1','rep2'), 3),
  dataset=c(rep('a',4), rep('b',2)),
  subgroup_1=c('x','x','y','y', 'x','y'), stringsAsFactors=FALSE)
params <- get_parameters(df, params_df)
params

#change specific arguments afterwards
params$a$calcMean <- FALSE
params$b$log2Transform <- TRUE

}
