% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{get_plot_options}
\alias{get_plot_options}
\title{Options from Excel}
\usage{
get_plot_options(options)
}
\arguments{
\item{options}{a data.frame}
}
\value{
Named list of values.
}
\description{
Reads Options from Excel template
}
\details{
options is a data.frame with columns named *Option* and *Value*,
  both case-sensitive. Several entries are vectors of character, numeric or
  boolean. This function tries to parse this correctly.
}
\examples{
  xl_fname <- system.file('extdata', 'example_excel_template.xls', package='seqNdisplayR')
  get_plot_options(xl_fname)

}
