% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{get_samples}
\alias{get_samples}
\title{Get Samples}
\usage{
get_samples(filled_df)
}
\arguments{
\item{filled_df}{a fill data frame (see details)}
}
\value{
Named lists or nested lists of named lists
}
\description{
Get Samples
}
\details{
Get samples from a data frame containing at a minimum columns
  dataset. Columns on the right to dataset may contain subcategories,
  called subgroup_1, subgroup_2 etc.
}
\examples{
df <- data.frame(dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), stringsAsFactors=FALSE)
get_samples(df)
df <- data.frame(dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), subgroup_2=c('a', 'b', 'a', 'b', NA, NA), stringsAsFactors=FALSE)
get_samples(df)

}
