% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{GetAnnotations}
\alias{GetAnnotations}
\title{Get Annotations}
\usage{
GetAnnotations(annotations)
}
\arguments{
\item{annotations}{Dataframe based on ANNOTATIONS sheet in Excel template}
}
\value{
Named list of path, or named list of GRanges. Names used are entries in the *annotation_name* column.
}
\description{
Internal function: 
Reads Annotations (paths and options) from Excel template
}
\details{
annotations is a data.frame with at least 2 columns *annotation_file* and *annotation_name*, both
  case-sensitive and the space is required.
}
\examples{
NULL

}
\author{
MS (minor additions by SLA)
}
\keyword{internal}
