% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{GetColors}
\alias{GetColors}
\title{Get Colors}
\usage{
GetColors(filled_df)
}
\arguments{
\item{filled_df}{a filled dataframe (see details)}
}
\value{
Named lists or nested lists of named lists
}
\description{
Internal function: 
Get colors from a data frame containing at a minimum columns color and dataset
}
\details{
Get colors from a data frame containing at a minimum columns
  color and dataset. Columns on the right to dataset may contain subcategories,
  called subgroup_1, subgroup_2 etc. UPS: dataset and subgroup_s must be the right-most columns, see example below.
}
\examples{
df = data.frame(color=c(rep('red', 4), rep('green', 2)),dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), subgroup_2=1:6, stringsAsFactors=FALSE)
GetColors(df)

}
\author{
MS
}
\keyword{From}
\keyword{GetColors}
\keyword{a}
\keyword{and}
\keyword{associated}
\keyword{based}
\keyword{chatGPT:}
\keyword{color}
\keyword{colors}
\keyword{column(s)}
\keyword{containing}
\keyword{corresponding}
\keyword{data}
\keyword{data.}
\keyword{extracts}
\keyword{filled_df}
\keyword{frame}
\keyword{from}
\keyword{function}
\keyword{in}
\keyword{information}
\keyword{internal}
\keyword{list}
\keyword{nested}
\keyword{on}
\keyword{or}
\keyword{organizes}
\keyword{returns}
\keyword{specific}
\keyword{the}
\keyword{values}
\keyword{vector}
\keyword{with}
