% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{GetPlotOptions}
\alias{GetPlotOptions}
\title{Get Plot Options}
\usage{
GetPlotOptions(options)
}
\arguments{
\item{options}{Dataframe based on GLOBAL_OPTIONS sheet in Excel template}
}
\value{
Named list of values
}
\description{
Internal function: 
Reads Global Options from Excel template
}
\details{
options is a data.frame with columns named *Option* and *Value*,
  both case-sensitive. Several entries are vectors of character, numeric or
  boolean. This function tries to parse this correctly.
}
\examples{
NULL

}
\author{
MS
}
\keyword{internal}
