% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{GetSamples}
\alias{GetSamples}
\title{Get Samples}
\usage{
GetSamples(filled_df)
}
\arguments{
\item{filled_df}{a filled dataframe (see details)}
}
\value{
Named lists or nested lists of named lists
}
\description{
Internal function: 
Get samples from a data frame containing at a minimum columns dataset
}
\details{
Get samples from a data frame containing at a minimum columns
  dataset. Columns on the right to dataset may contain subcategories,
  called subgroup_1, subgroup_2 etc.
}
\examples{
df = data.frame(dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), stringsAsFactors=FALSE)
GetSamples(df)
df = data.frame(dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), subgroup_2=c('a', 'b', 'a', 'b', NA, NA), stringsAsFactors=FALSE)
GetSamples(df)

}
\author{
MS
}
\keyword{From}
\keyword{GetSamples}
\keyword{a}
\keyword{and}
\keyword{associated}
\keyword{based}
\keyword{character}
\keyword{chatGPT:}
\keyword{column(s)}
\keyword{containing}
\keyword{corresponding}
\keyword{data}
\keyword{data.}
\keyword{extracts}
\keyword{filled_df}
\keyword{frame}
\keyword{from}
\keyword{function}
\keyword{in}
\keyword{information}
\keyword{internal}
\keyword{list}
\keyword{nested}
\keyword{on}
\keyword{or}
\keyword{organizes}
\keyword{returns}
\keyword{sample}
\keyword{samples}
\keyword{specific}
\keyword{the}
\keyword{unique}
\keyword{values}
\keyword{vector}
\keyword{with}
