% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayRSession.R
\name{glimpse_session}
\alias{glimpse_session}
\title{Glimpse session}
\usage{
glimpse_session(samples, colors, bigwigs, level = 0, indent_size = "   ")
}
\arguments{
\item{samples}{samples object as used by seqNdisplayR}

\item{colors}{colors object as used by seqNdisplayR}

\item{bigwigs}{bigwigs object as used by seqNdisplayR}

\item{indent_size}{string used for indent spacing of levels in the outpur}

\item{levels}{used only for internal recursion, don't change default=1.}
}
\value{
Print to R session.
}
\description{
Prints an overview over samples, colors and associated bigwigs in a seqNdisplayR Session.
}
\details{
Convenience function for checking parsing of samples, colors and bigwigs.
}
\examples{

}
