% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{Hex2Hsl}
\alias{Hex2Hsl}
\title{Hex2Hsl}
\usage{
Hex2Hsl(c)
}
\arguments{
\item{c}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
Hex2Hsl('black')
Hex2Hsl('blue')
Hex2Hsl('#FFA550')

}
\author{
SLA (based on https://bitbucket.org/mattgove/color-theory/src/master/mg_colors.py)
}
\keyword{internal}
