% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{is.color}
\alias{is.color}
\title{is.color}
\usage{
is.color(put_color_vector)
}
\description{
is.color
}
\details{
Internal function: Helper function that determines whether input (vector) is a color
}
