% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsEmpty}
\alias{IsEmpty}
\title{Is Empty}
\usage{
IsEmpty(x)
}
\arguments{
\item{x}{vector}
}
\value{
TRUE/FALSE for each element in input vector
}
\description{
Internal function: 
Convenience function checks whether all NA or empty string
}
\examples{
IsEmpty(NA)
IsEmpty(c(NA, NA))
IsEmpty(c('', NA))
IsEmpty(c('fnyg', NA)) 

}
\author{
MS
}
\keyword{internal}
