% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ListExamplesSampleSheets}
\alias{ListExamplesSampleSheets}
\title{List Examples Sample Sheets}
\usage{
ListExamplesSampleSheets()
}
\value{
the example seqNdisplayR Excel Templates that comes with the package
}
\description{
Finds and outputs the example seqNdisplayR Excel Templates that comes with the package
}
\examples{
example_sample_sheets = ListExamplesSampleSheets()
example_folder = ExamplesSampleSheetsFolder()
xl_fname = paste0(example_folder, example_sample_sheets[1])
session = LoadExcel(xl_fname, load_annotations = T)
plot(session, feature='LMO4')

}
\author{
SLA
}
