% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ListFlatten}
\alias{ListFlatten}
\title{Flatten Nested List}
\usage{
ListFlatten(x, prefix = "")
}
\arguments{
\item{x}{a (nested) list}

\item{prefix}{internal parameter for recursion (do not set manually)}
}
\value{
a flat list with concatenated names
}
\description{
Internal function:
Recursively flatten a nested list into a single-level named list.
Base R replacement for rlist::list.flatten.
Names are concatenated with "." to preserve hierarchy
(e.g. list(a = list(b = 1)) becomes list("a.b" = 1)).
}
\examples{
l <- list(a = list(b = 1, c = 2), d = 3)
ListFlatten(l)
# $a.b  [1] 1
# $a.c  [1] 2
# $d    [1] 3

}
\author{
SLA
}
\keyword{internal}
