% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{load_excel}
\alias{load_excel}
\title{Load Excel Template}
\usage{
load_excel(xl_fname, load_annotations = FALSE)
}
\arguments{
\item{load_annotations}{load annotations as GRanges? default=FALSE.}

\item{fname}{path to an excel template file}
}
\value{
seqNdisplayRSession object, essentially a named list with slots samples,
  colors, bigwig_dirs, bigwigs, parameters and annotations.
}
\description{
Load Excel Template
}
\details{
Load Excel template and parses information to seqNdisplayR session
  object. See Excel template sheet in
  \code{inst/extdata/example_excel_template.xls} for more information.
}
