% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{LoadTracks}
\alias{LoadTracks}
\title{Load Tracks}
\usage{
LoadTracks(
  plotted_region,
  samples,
  bigwigs,
  bigwig_dirs,
  parameters,
  verbosity
)
}
\arguments{
\item{plotted_region}{}

\item{samples}{}

\item{bigwigs}{}

\item{bigwig_dirs}{}

\item{parameters}{}

\item{verbosity}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
