% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{next_grpcol}
\alias{next_grpcol}
\title{Get next group column}
\usage{
next_grpcol(df, col)
}
\arguments{
\item{df}{data.frame}

\item{col}{a column index}
}
\description{
Get next group column
}
\details{
Returns ncol+1 if no next column found
}
\examples{
  df <- data.frame(dataset=c(rep('a',3),rep('b',2)), subgroup_1 = c('a', 'b', 'b', 'a', 'b'), subgroup_2 = c('a', 'a', 'b', '', ''))
  next_grpcol(df, 1)
  next_grpcol(df, 2)
  next_grpcol(df, 3)
  df <- data.frame(dataset=c(rep('a',3),rep('b',2)), subgroup_2 = c('a', 'b', 'b', 'a', 'b'), subgroup_1 = c('a', 'a', 'b', '', ''))
  next_grpcol(df, 1)
  next_grpcol(df, 2)
  next_grpcol(df, 3)

}
