% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NumberingSpacers}
\alias{NumberingSpacers}
\title{Numbering Spacers}
\usage{
NumberingSpacers(stranded_list)
}
\arguments{
\item{stranded_list}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
