% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{ordered_split}
\alias{ordered_split}
\title{Ordered Split}
\usage{
ordered_split(x, f, drop = TRUE)
}
\description{
Ordered Split
}
\details{
Split data frame by values in a column but maintain order as order of
  appearance. ie different to base R split.data.frame where order of split
  objects are ordered as default factors.
}
