% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrganizeAllAnnotationTextsInPlottedRegion}
\alias{OrganizeAllAnnotationTextsInPlottedRegion}
\title{Organize All Annotation Texts In Plotted Region}
\usage{
OrganizeAllAnnotationTextsInPlottedRegion(
  feature_text,
  plotted_region,
  letter_widths
)
}
\arguments{
\item{feature_text}{}

\item{plotted_region}{}

\item{letter_widths}{}
}
\value{
placeholder
}
\description{
Internal function: 
Determine the y-axis organization of names of annotated features within plotted region depending on font size
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
