% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{OrganizeAnnotatedFeaturesInRegion}
\alias{OrganizeAnnotatedFeaturesInRegion}
\title{Organize Annotated Features In Region}
\usage{
OrganizeAnnotatedFeaturesInRegion(
  plotted_region,
  annotations,
  incl_feature_names = NULL,
  center_of_mass = FALSE,
  incl_feature_brackets = NULL,
  annotation_packing = NULL
)
}
\arguments{
\item{plotted_region}{}

\item{annotations}{}
}
\value{
placeholder
}
\description{
Internal function: 
Wrapper function that executes the three above functions
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
