% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrganizeAnnotationText}
\alias{OrganizeAnnotationText}
\title{Organize Annotation Text}
\usage{
OrganizeAnnotationText(
  plotted_region,
  annot_info,
  annotation_packing,
  letter_widths,
  center_of_mass = FALSE,
  verbosity
)
}
\arguments{
\item{plotted_region}{}

\item{annot_info}{}

\item{annotation_packing}{}

\item{letter_widths}{}

\item{center_of_mass}{}

\item{verbosity}{}
}
\value{
placeholder
}
\description{
Internal function: 
Given a set of possible font sizes (e.g. helvetica 4-8 given by letter_widths), where should the feature names within the region be plotted
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
