% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{OrganizedPanelsList}
\alias{OrganizedPanelsList}
\title{Organized Panels List}
\usage{
OrganizedPanelsList(tracks_listed)
}
\arguments{
\item{tracks_listed}{}
}
\value{
placeholder
}
\description{
Internal function: 
Organize the panels with track-names to be plotted on the left side of the tracks
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
