% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotAnnotation}
\alias{PlotAnnotation}
\title{Plot Annotation}
\usage{
PlotAnnotation(
  annot_info,
  stranded,
  annot_cols,
  annotation_packing,
  plotted_region,
  plotted_strand,
  substrand,
  basic_plot_parameters,
  plot_start,
  plot_end,
  plot_width,
  bin_size,
  reverse_strand_direction,
  incl_feature_names,
  feature_names_above,
  incl_feature_brackets,
  incl_feature_shadings,
  feature_shading_colors,
  feature_shading_alpha,
  plot_width_parameters,
  plot_vertical_parameters,
  windows_height,
  feature_font_size,
  annotation_panel_font_size,
  annot_panel_dist = 0.4,
  coords_tracks,
  font_colors,
  font_family,
  first_plot,
  scaling_factor,
  verbosity
)
}
\arguments{
\item{annot_info}{}

\item{stranded}{}

\item{annot_cols}{}

\item{annotation_packing}{}

\item{plotted_region}{}

\item{plotted_strand}{}

\item{substrand}{}

\item{basic_plot_parameters}{}

\item{plot_start}{}

\item{plot_end}{}

\item{plot_width}{}

\item{bin_size}{}

\item{reverse_strand_direction}{}

\item{incl_feature_names}{}

\item{feature_names_above}{}

\item{incl_feature_brackets}{}

\item{incl_feature_shadings}{}

\item{feature_shading_colors}{}

\item{feature_shading_alpha}{}

\item{annotation_panel_font_size}{}

\item{annot_panel_dist}{}

\item{coords_tracks}{}

\item{font_colors}{}

\item{font_family}{}

\item{first_plot}{}

\item{scaling_factor}{}

\item{verbosity}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
