% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotHeader}
\alias{PlotHeader}
\title{Plot Header}
\usage{
PlotHeader(
  windows_height,
  n_segment,
  coords_tracks,
  full_width_cm,
  plot_width,
  header,
  header_font_sizes,
  chrom,
  both_strands,
  plotted_strand,
  plot_start,
  plot_end,
  font_colors,
  font_family,
  first_plot,
  scaling_factor
)
}
\arguments{
\item{windows_height}{}

\item{n_segment}{}

\item{coords_tracks}{}

\item{full_width_cm}{}

\item{plot_width}{}

\item{header}{}

\item{header_font_sizes}{}

\item{chrom}{}

\item{both_strands}{}

\item{plotted_strand}{}

\item{plot_start}{}

\item{plot_end}{}

\item{font_colors}{}

\item{font_family}{}

\item{first_plot}{}

\item{scaling_factor}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
