% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PlotHeightParameters}
\alias{PlotHeightParameters}
\title{Plot Height Parameters}
\usage{
PlotHeightParameters(
  combined_track_vector,
  track_vector,
  annotation_lines,
  total_annotation_lines,
  annot_heights_incl_text,
  max_annot_lines,
  annot_heights,
  track_height_cm,
  full_height_cm,
  title_field_height_cm,
  genomic_scale_height_cm,
  annotation_height_cm,
  spacer_height_cm,
  plot_vertical_parameters
)
}
\arguments{
\item{combined_track_vector}{}

\item{track_vector}{}

\item{annotation_lines}{}

\item{total_annotation_lines}{}

\item{annot_heights_incl_text}{}

\item{max_annot_lines}{}

\item{annot_heights}{}

\item{track_height_cm}{}

\item{full_height_cm}{}

\item{title_field_height_cm}{}

\item{genomic_scale_height_cm}{}

\item{annotation_height_cm}{}

\item{spacer_height_cm}{}

\item{plot_vertical_parameters}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
