% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotHeightParameters_obs}
\alias{PlotHeightParameters_obs}
\title{Plot Height Parameters}
\usage{
PlotHeightParameters_obs(
  annot_info,
  track_vector,
  max_annot_lines,
  annot_heights,
  letter_heights,
  incl_feature_names,
  feature_text_org,
  annotation_packing,
  incl_feature_brackets,
  track_height_cm,
  full_height_cm,
  stranded_beds,
  plot_vertical_parameters
)
}
\description{
Plot Height Parameters
}
\details{
Internal function:
}
