% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotPanels}
\alias{PlotPanels}
\title{Plot Panels}
\usage{
PlotPanels(
  plotting_segment,
  plotted_strand,
  panel_info,
  panels_list,
  panel_separators,
  separators_lwds,
  separators_colors,
  incl_first_panel,
  print_one_line_sample_names,
  replicate_names,
  plot_width_parameters,
  windows_height,
  vertical_slots,
  segment_top,
  full_width_cm,
  font_color,
  font_family,
  colors,
  first_plot,
  letter_heights,
  scaling_factor
)
}
\arguments{
\item{plotting_segment}{}

\item{plotted_strand}{}

\item{panel_info}{}

\item{panels_list}{}

\item{panel_separators}{}

\item{separators_lwds}{}

\item{separators_colors}{}

\item{incl_first_panel}{}

\item{print_one_line_sample_names}{}

\item{replicate_names}{}

\item{plot_width_parameters}{}

\item{windows_height}{}

\item{vertical_slots}{}

\item{segment_top}{}

\item{full_width_cm}{}

\item{font_color}{}

\item{font_family}{}

\item{colors}{}

\item{first_plot}{}

\item{letter_heights}{}

\item{scaling_factor}{}
}
\value{
placeholder
}
\description{
Internal function: 
Sample info at the left side of the plots
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
