% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSpacer}
\alias{PlotSpacer}
\title{Plot Spacer}
\usage{
PlotSpacer(
  windows_height,
  spacer_segment,
  right_border,
  plotted_strand,
  neg_vals_neg_strand,
  panel_separators,
  separators_lwds,
  separators_colors,
  scaling_factor
)
}
\arguments{
\item{windows_height}{}

\item{spacer_segment}{}

\item{right_border}{}

\item{plotted_strand}{}

\item{neg_vals_neg_strand}{}

\item{panel_separators}{}

\item{separators_lwds}{}

\item{separators_colors}{}

\item{scaling_factor}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
