% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{PlotVerticalParameters}
\alias{PlotVerticalParameters}
\title{Plot Vertical Parameters}
\usage{
PlotVerticalParameters()
}
\value{
a named vector with default vertical proportions of various plotting segment types relative to 'seq' tracks
}
\description{
Internal function
}
\examples{
plot_vertical_parameters = PlotVerticalParameters()

}
\author{
SLA
}
\keyword{internal}
