% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RbindFill}
\alias{RbindFill}
\title{Bind Rows with Fill}
\usage{
RbindFill(df_list)
}
\arguments{
\item{df_list}{a list of data frames}
}
\value{
a single data frame
}
\description{
Internal function:
Bind a list of data frames by row, filling missing columns with NA.
Base R replacement for dplyr::bind_rows.
}
\keyword{internal}
