% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{RelativeAnnotationHeight}
\alias{RelativeAnnotationHeight}
\title{RelativeAnnotationHeight}
\usage{
RelativeAnnotationHeight(
  annot_info,
  annot_heights,
  letter_heights,
  incl_feature_names,
  annotation_packing,
  incl_feature_brackets,
  stranded_beds
)
}
\arguments{
\item{annot_info}{}

\item{annot_heights}{}

\item{stranded_beds}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
