% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Session2xlsx}
\alias{Session2xlsx}
\title{Session2xlsx}
\usage{
Session2xlsx(session, path, ...)
}
\arguments{
\item{session}{seqNdisplayRSession object}

\item{path}{excel file name including full path}

\item{...}{}
}
\value{
Write Excel template as described in \link{LoadExcel}.
}
\description{
Save session object to Excel xlsx-file
}
\examples{
NULL

}
\author{
MS (minor additions by SLA)
}
