% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayRSession.R
\name{session_to_xlsx}
\alias{session_to_xlsx}
\title{Save Session To Excel}
\usage{
session_to_xlsx(session, path, ...)
}
\arguments{
\item{session}{seqNdisplayRSession object}

\item{path}{Excel file path}

\item{strand_regex}{named vector c('+': ..., '-': ...) for regex for converting plus strand to minus strand bigwig names.}
}
\value{
nothing
}
\description{
Save Session To Excel
}
\details{
Write Excel template as described in \link{load_excel}.
}
\examples{

}
