% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sNdR_clear_log}
\alias{sNdR_clear_log}
\title{Clear seqNdisplayR Message Log}
\usage{
sNdR_clear_log()
}
\description{
Developer tool: clear the internal message log.
}
\examples{
\dontrun{
sNdR_clear_log()
}

}
