% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{UpdatePlotVerticalParameters}
\alias{UpdatePlotVerticalParameters}
\title{Update Plot Vertical Parameters}
\usage{
UpdatePlotVerticalParameters(
  plot_vertical_parameters,
  track_height_cm_estimate,
  title_field_height_cm,
  genomic_scale_height_cm,
  annotation_height_cm,
  spacer_height_cm
)
}
\arguments{
\item{plot_vertical_parameters}{}

\item{track_height_cm_estimate}{}

\item{title_field_height_cm}{}

\item{genomic_scale_height_cm}{}

\item{annotation_height_cm}{}

\item{spacer_height_cm}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
