% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AdjustColorPhi}
\alias{AdjustColorPhi}
\title{Adjust Color Phi}
\usage{
AdjustColorPhi(c_hsl, phi = 180)
}
\arguments{
\item{c_hsl}{}

\item{phi}{}
}
\value{

}
\description{
Internal function:
}
\note{
change throughout from adjust_color_phi to AdjustColorPhi - done? #% 2022-10-04
}
\examples{
c_hsl = Hex2Hsl('#FFA550')
c_hsl_180 = AdjustColorPhi(c_hsl)
 
}
\author{
SLA (based on https://bitbucket.org/mattgove/color-theory/src/master/mg_colors.py)
}
