% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AdjustEstimatedPlotHeights}
\alias{AdjustEstimatedPlotHeights}
\title{Adjust Estimated Plot Heights}
\usage{
AdjustEstimatedPlotHeights(
  estimated_plot_heights,
  plot_vertical_parameters,
  full_height_cm,
  track_height_cm,
  title_field_height_cm,
  genomic_scale_height_cm,
  annotation_height_cm,
  spacer_height_cm
)
}
\arguments{
\item{estimated_plot_heights}{}

\item{plot_vertical_parameters}{}

\item{full_height_cm}{}

\item{track_height_cm}{}

\item{title_field_height_cm}{}

\item{genomic_scale_height_cm}{}

\item{annotation_height_cm}{}

\item{spacer_height_cm}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
