% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AlignBasicPlotParameters}
\alias{AlignBasicPlotParameters}
\title{Align Basic Plot Parameters}
\usage{
AlignBasicPlotParameters(
  basic_plot_parameters,
  both_strands,
  strands_intermingled,
  fixed_plot_vertical_parameters
)
}
\arguments{
\item{basic_plot_parameters}{}

\item{both_strands}{}

\item{strands_intermingled}{}

\item{fixed_plot_vertical_parameters}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
\keyword{internal}
