% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AlignBasicPlotParameters}
\alias{AlignBasicPlotParameters}
\title{Align Basic Plot Parameters}
\usage{
AlignBasicPlotParameters(
  basic_plot_parameters,
  both_strands,
  strands_intermingled,
  fixed_plot_vertical_parameters
)
}
\description{
Align Basic Plot Parameters
}
\details{
Internal function:
}
