% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AllEmpty}
\alias{AllEmpty}
\title{All Empty}
\usage{
AllEmpty(x)
}
\arguments{
\item{x}{vector}
}
\value{
TRUE/FALSE
}
\description{
Internal function: 
Convenience function checks whether all NA or empty in a vector of strings
}
\note{
change throughout from allempty to AllEmpty - done? #% 2022-10-04
}
\examples{
AllEmpty(NA)
AllEmpty(c(NA, NA))
AllEmpty(c('', NA))
AllEmpty(c('fnyg', NA)) 

}
\author{
MS
}
