% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{AnnotatedFeaturesInRegion}
\alias{AnnotatedFeaturesInRegion}
\title{Annotated Features In Region}
\usage{
AnnotatedFeaturesInRegion(plotted_region, annotations)
}
\arguments{
\item{plotted_region}{}

\item{annotations}{}
}
\description{
Internal function: 
This function is the first step in organizing the annotated features in the plotted region into various formats
Used together with "OrganizeOverlappingLoci" and "ConvertCollapsedFormat" to create a final organized version of the annotated features
}
\author{
SLA
}
\keyword{internal}
