% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{BasicPlotParameters}
\alias{BasicPlotParameters}
\title{Basic Plot Parameters}
\usage{
BasicPlotParameters(
  plotted_strand,
  plotted_region,
  feature_names_font_size,
  plot_height_parameters,
  plot_width_parameters,
  full_width_cm,
  full_height_cm = NULL,
  track_height_cm = 0.3,
  plot_vertical_parameters,
  bin_size,
  bins_per_cm,
  plotting_segment_order,
  tracks_listed,
  unstranded_beds
)
}
\description{
Basic Plot Parameters
}
\details{
Internal function:
}
