% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{BuildScrutinizePlotSegmentOrder}
\alias{BuildScrutinizePlotSegmentOrder}
\title{Build Scrutinize Plot Segment Order}
\usage{
BuildScrutinizePlotSegmentOrder(
  plotting_segment_order,
  plotted_region,
  datasets,
  plotted_samples,
  header,
  include_genomic_scale,
  genomic_scale_on_top,
  incl_annot,
  horizontal_spacers,
  tracks_listed,
  both_strands,
  any_stranded_beds,
  any_unstranded_beds,
  strands_intermingled,
  verbosity,
  interface
)
}
\arguments{
\item{plotting_segment_order}{}

\item{plotted_region}{}

\item{datasets}{}

\item{plotted_samples}{}

\item{header}{}

\item{include_genomic_scale}{}

\item{genomic_scale_on_top}{}

\item{incl_annot}{}

\item{horizontal_spacers}{}

\item{tracks_listed}{}

\item{both_strands}{}

\item{any_stranded_beds}{}

\item{any_unstranded_beds}{}

\item{strands_intermingled}{}

\item{verbosity}{}

\item{interface}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
\keyword{internal}
