% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{BuildScrutinizePlotSegmentOrder}
\alias{BuildScrutinizePlotSegmentOrder}
\title{BuildScrutinizePlotSegmentOrder}
\usage{
BuildScrutinizePlotSegmentOrder(
  plotting_segment_order,
  plotted_region,
  datasets,
  plotted_samples,
  header,
  include_genomic_scale,
  genomic_scale_on_top,
  incl_annot,
  horizontal_spacers,
  tracks_listed,
  both_strands,
  any_stranded_beds,
  any_unstranded_beds,
  strands_intermingled,
  verbosity,
  interface
)
}
\description{
BuildScrutinizePlotSegmentOrder
}
\details{
Internal function:
}
