% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{CalculateTrackHeight}
\alias{CalculateTrackHeight}
\title{Update Plot Vertical Parameters v2}
\usage{
CalculateTrackHeight(
  combined_track_vector,
  total_annotation_lines,
  full_height_cm,
  title_field_height_cm,
  genomic_scale_height_cm,
  annotation_height_cm,
  spacer_height_cm
)
}
\description{
Update Plot Vertical Parameters v2
}
\details{
Internal function:
}
