% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{CalculateTrackHeight}
\alias{CalculateTrackHeight}
\title{Calculate Track Height}
\usage{
CalculateTrackHeight(
  combined_track_vector,
  total_annotation_lines,
  full_height_cm,
  title_field_height_cm,
  genomic_scale_height_cm,
  annotation_height_cm,
  spacer_height_cm
)
}
\arguments{
\item{combined_track_vector}{}

\item{total_annotation_lines}{}

\item{full_height_cm}{}

\item{title_field_height_cm}{}

\item{genomic_scale_height_cm}{}

\item{annotation_height_cm}{}

\item{spacer_height_cm}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
