% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{CommonPrefix}
\alias{CommonPrefix}
\title{Common Prefix}
\usage{
CommonPrefix(vector_of_strings)
}
\arguments{
\item{vector_of_strings}{}
}
\value{
A single string which is a common prefix in all strings
}
\description{
Internal function: 
Find common prefix
}
\examples{

}
\author{
MS
}
\keyword{internal}
