% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{ConstantsDefaults}
\alias{ConstantsDefaults}
\title{Constants and Defaults}
\usage{
ConstantsDefaults()
}
\value{
a named vector with constants and defaults (package-wide available arguments)
}
\description{
Internal function
Container for Constants and Defaults
}
\note{
implement function throughout - done? #% 2022-10-05
}
\examples{
constants_defaults = ConstantsDefaults()
line_width_scaling_factor = constants_defaults['line_width_scaling_factor']
points_per_cm = constants_defaults['points_per_cm']
arrow_constant = constants_defaults['arrow_constant']
cm_to_in = constants_defaults['cm_to_in']
std_letter_width = constants_defaults['std_letter_width']
std_letter_height = constants_defaults['std_letter_height']
min_font_size = constants_defaults['min_font_size']
annot_panel_dist = constants_defaults['annot_panel_dist']

}
\author{
SLA
}
