% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{CoordsOfGenomicScale}
\alias{CoordsOfGenomicScale}
\title{Coords Of Genomic Scale}
\usage{
CoordsOfGenomicScale(
  chrom_coord,
  font_size,
  coords_tracks,
  full_width_cm,
  plot_width,
  plot_start,
  plot_end,
  letter_widths
)
}
\description{
Coords Of Genomic Scale
}
\details{
Internal function:
}
