% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{DefaultAnnotationOptions}
\alias{DefaultAnnotationOptions}
\title{Default Annotation Options}
\usage{
DefaultAnnotationOptions()
}
\value{
a named list of default annotation-display options.
}
\description{
Internal function
}
\note{
change throughout to DefaultAnnotationOptions - done? #% 2022-10-05
}
\examples{
default_annotation_options = DefaultAnnotationOptions()

}
\author{
MS/SLA
}
