% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{DefaultParameters}
\alias{DefaultParameters}
\title{Default Parameters}
\usage{
DefaultParameters()
}
\value{
a named list of default Dataset-specific options
}
\description{
Internal function
}
\examples{
default_parameters = DefaultParameters()

}
\author{
MS/SLA
}
\keyword{internal}
