% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{DefaultPlotOptions}
\alias{DefaultPlotOptions}
\title{Default Plot Options}
\usage{
DefaultPlotOptions()
}
\value{
a named list of default plot options
}
\description{
Internal function
}
\note{
change throughout to DefaultPlotOptions - done? #% 2022-10-05
}
\examples{
default_plot_options = DefaultPlotOptions()

}
\author{
MS/SLA
}
