% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{DeleteNULLs}
\alias{DeleteNULLs}
\title{Delete NULLs}
\usage{
DeleteNULLs(x_list)
}
\arguments{
\item{x_list}{}
}
\value{

}
\description{
Internal function: 
Delele null/empty entries in a list
}
\examples{

}
\author{
SLA
}
\keyword{internal}
