% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{Empty2Null}
\alias{Empty2Null}
\title{Empty 2 Null}
\usage{
Empty2Null(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Internal function: 
Helper used for parsing parameters$<dataset>$whichSamples
}
\examples{

}
\author{
MS
}
\keyword{internal}
