% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{EstimatePlotHeights}
\alias{EstimatePlotHeights}
\title{Estimate Plot Heights}
\usage{
EstimatePlotHeights(
  annot_info,
  incl_feature_names,
  annotation_packing,
  incl_feature_brackets,
  plotting_segment_order,
  tracks_listed,
  track_height_cm,
  full_height_cm,
  stranded_beds,
  plot_vertical_parameters,
  verbosity,
  interface
)
}
\description{
Estimate Plot Heights
}
\details{
Internal function:
}
