% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{EstimatePlotHeights}
\alias{EstimatePlotHeights}
\title{Estimate Plot Heights}
\usage{
EstimatePlotHeights(
  annot_info,
  incl_feature_names,
  annotation_packing,
  incl_feature_brackets,
  plotting_segment_order,
  tracks_listed,
  track_height_cm,
  full_height_cm,
  stranded_beds,
  plot_vertical_parameters,
  verbosity,
  interface
)
}
\arguments{
\item{annot_info}{}

\item{incl_feature_names}{}

\item{annotation_packing}{}

\item{incl_feature_brackets}{}

\item{plotting_segment_order}{}

\item{tracks_listed}{}

\item{track_height_cm}{}

\item{full_height_cm}{}

\item{stranded_beds}{}

\item{plot_vertical_parameters}{}

\item{verbosity}{}

\item{interface}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
