% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{EvaluateNumericValue}
\alias{EvaluateNumericValue}
\title{Evaluate Numeric Value}
\usage{
EvaluateNumericValue(
  num_val,
  positive_val,
  min_val,
  max_val,
  interval_obligatory = FALSE,
  turn_errors_to_warnings = FALSE,
  alt_par_name = NULL,
  verbosity
)
}
\description{
Evaluate Numeric Value
}
\details{
Internal function: Evaluate whether a vector/value is numeric and falls within recommmended range
}
