% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{FillDf}
\alias{FillDf}
\title{Fill Df}
\usage{
FillDf(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Dataframe
}
\description{
Internal function: 
Fill Empty Rows in individual columns in dataframe
}
\details{
Empty rows in dataframe from loaded XL sheet are filled with values
  from columns above in hierarchical fashion such that the right-most columns
  are only filled if specified in the dataset defining row of the seqtype.
  See Excel template sheet in \code{inst/extdata/example_excel_template.xls}
  for more information.
}
\examples{
NULL

}
\author{
MS/SLA
}
\keyword{internal}
