% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{FinalizePanelsDimensions}
\alias{FinalizePanelsDimensions}
\title{Finalize Panels Dimensions}
\usage{
FinalizePanelsDimensions(panel_info, both_strands)
}
\arguments{
\item{panel_info}{}

\item{both_strands}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
