% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{FinalizePanelsDimensions}
\alias{FinalizePanelsDimensions}
\title{Finalize Panels Dimensions}
\usage{
FinalizePanelsDimensions(panel_info, both_strands)
}
\description{
Finalize Panels Dimensions
}
\details{
Internal function:
}
