% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{FinalizePlottingSegmentOrder}
\alias{FinalizePlottingSegmentOrder}
\title{Finalize Plotting Segment Order}
\usage{
FinalizePlottingSegmentOrder(
  plotting_segment_order,
  tracks_listed,
  both_strands,
  include_genomic_scale,
  genomic_scale_on_top,
  any_stranded_beds,
  any_unstranded_beds,
  strands_intermingled
)
}
\description{
Finalize Plotting Segment Order
}
\details{
Internal function:
}
