% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{FinalOrganizedAnnotationTextsInPlottedRegion}
\alias{FinalOrganizedAnnotationTextsInPlottedRegion}
\title{Final Organized Annotation Texts In Plotted Region}
\usage{
FinalOrganizedAnnotationTextsInPlottedRegion(
  organized_annotation_texts,
  annotation_names,
  font_size
)
}
\arguments{
\item{organized_annotation_texts}{}

\item{annotation_names}{}

\item{font_size}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
