% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{ForceScaleList}
\alias{ForceScaleList}
\title{Force Scale List}
\usage{
ForceScaleList(samples, force_scale, strands, verbosity, interface)
}
\arguments{
\item{samples}{}

\item{force_scale}{}

\item{strands}{}

\item{verbosity}{}

\item{interface}{}
}
\description{
Internal function: 
Checks/constructs a force_scale_list
}
\author{
SLA
}
\keyword{internal}
