% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{GetBigwigDirs}
\alias{GetBigwigDirs}
\title{Get Bigwig Dirs}
\usage{
GetBigwigDirs(filled_df)
}
\arguments{
\item{filled_df}{a filled dataframe (see details)}
}
\value{
Named vector
}
\description{
Internal function: 
Get bigwig dirs file names from a data frame containing at a minimum columns bigwig_file and dataset
}
\details{
Get bigwig dirs file names from a data frame containing at a minimum columns
  bigwig_file and dataset. Note: in seqNdisplayR only one bigwig_dir is allowed per seqtype.
}
\note{
change throughout from get_bigwig_dirs to GetBigwigDirs - done? #% 2022-10-04
}
\examples{
df = data.frame(bigwig_directory=c(rep('http://seqA/', 4), rep('http://seqB/', 2)),dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), subgroup_2=1:6, stringsAsFactors=FALSE)
GetBigwigDirs(df)

}
\author{
MS
}
