% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{GetBigwigs}
\alias{GetBigwigs}
\title{Get Bigwigs}
\usage{
GetBigwigs(filled_df)
}
\arguments{
\item{filled_df}{a filled dataframe (see details)}
}
\value{
Named lists or nested lists of named lists
}
\description{
Internal function: 
Get bigwig file names from a data frame containing at a minimum columns bigwig_file, strand and dataset
}
\details{
Get bigwig file names from a data frame containing at a minimum columns
  bigwig_file, strand and dataset. Strand must be 'plus' or 'minus', use 'plus' for unstranded data. Columns on the right to dataset may contain subcategories,
  called subgroup_1, subgroup_2 etc. UPS: dataset and subgroup_s must be the right-most columns, see example below.
}
\note{
change throughout from get_bigwigs to GetBigwigs - done? #% 2022-10-04
}
\examples{
df = data.frame(bigwig_file=c('a.bw', 'b.bw', 'c.bw', 'd.bw', 'e.bw', 'f.bw'),
  strand = rep('plus', 6), dataset=c(rep('a',4), rep('b',2)),
  subgroup_1=c('x','x','y','y', 'x','y'),
  stringsAsFactors=FALSE)
GetBigwigs(df)

df = data.frame(
  bigwig_file=c('a_plus.bw', 'a_minus.bw', 'b_plus.bw', 'b_minus.bw', 'c_plus.bw', 'c_minus.bw'),
  strand = rep(c('plus', 'minus'), 3),
  dataset=c(rep('a',4), rep('b',2)),
  subgroup_1=c('x','x','y','y', 'x','y'),
  stringsAsFactors=FALSE)
GetBigwigs(df)

}
\author{
MS
}
