% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{GetBinSize}
\alias{GetBinSize}
\title{Get Bin Size}
\usage{
GetBinSize(bin_size, plot_width, tracks_width_cm, bins_per_cm, verbosity)
}
\arguments{
\item{bin_size}{}

\item{plot_width}{}

\item{tracks_width_cm}{}

\item{bins_per_cm}{}

\item{verbosity}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
