% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{GetSamples}
\alias{GetSamples}
\title{Get Samples}
\usage{
GetSamples(filled_df)
}
\arguments{
\item{filled_df}{a filled dataframe (see details)}
}
\value{
Named lists or nested lists of named lists
}
\description{
Internal function: 
Get samples from a data frame containing at a minimum columns dataset
}
\details{
Get samples from a data frame containing at a minimum columns
  dataset. Columns on the right to dataset may contain subcategories,
  called subgroup_1, subgroup_2 etc.
}
\note{
change throughout from get_samples to GetSamples - done? #% 2022-10-04
}
\examples{
df = data.frame(dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), stringsAsFactors=FALSE)
GetSamples(df)
df = data.frame(dataset=c(rep('a',4), rep('b',2)), subgroup_1=c('x','x','y','y', 'x','y'), subgroup_2=c('a', 'b', 'a', 'b', NA, NA), stringsAsFactors=FALSE)
GetSamples(df)

}
\author{
MS
}
