% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{GlimpseSession}
\alias{GlimpseSession}
\title{Glimpse Session}
\usage{
GlimpseSession(samples, colors, bigwigs, level = 0, indent_size = "   ")
}
\arguments{
\item{samples}{samples object as used by seqNdisplayR}

\item{colors}{colors object as used by seqNdisplayR}

\item{bigwigs}{bigwigs object as used by seqNdisplayR}

\item{indent_size}{string used for indent spacing of levels in the output}

\item{levels}{used only for internal recursion, don't change default=1.}
}
\value{
Print to R session
}
\description{
Internal function: 
Prints an overview over samples, colors and associated bigwigs in a seqNdisplayR Session.
}
\details{
Convenience function for checking parsing of samples, colors and bigwigs.
}
\author{
MS
}
\keyword{internal}
