% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{HandleForcedScaleFromParameters}
\alias{HandleForcedScaleFromParameters}
\title{Parse Option String}
\usage{
HandleForcedScaleFromParameters(pars)
}
\arguments{
\item{option_str}{string representation of the option}
}
\description{
Parse string into relevant R object class
}
\details{
String can represent a named list, unnamed list, named vector, unnamed vector or single value. If string contains ";" assumes a list; if string contains "," assumes a vector; individual strings are interprated as "NULL" -> NULL; if "TRUE" or "FALSE" --> TRUE/FALSE; if single number --> as.numeric; if single non-number --> as.character;

Force Scale List

Internal function: Checks/constructs a force_scale_list
handle force_scale which is part of parameters but needs to passed differently to plot function
}
