% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{HandleForcedScaleFromParameters}
\alias{HandleForcedScaleFromParameters}
\title{Handle Forced Scale From Parameters}
\usage{
HandleForcedScaleFromParameters(pars)
}
\arguments{
\item{pars}{}
}
\value{

}
\description{
Internal function: 
Checks/constructs a force_scale_list
Handle force_scale which is part of "parameters" argument but needs to passed differently to plot function
}
\examples{

}
\author{
MS/SLA
}
