% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{Hex2Hsl}
\alias{Hex2Hsl}
\title{Hex2Hsl}
\usage{
Hex2Hsl(c)
}
\arguments{
\item{c}{}
}
\value{

}
\description{
Internal function:
}
\note{
change throughout from hex2hsl to Hex2Hsl - done? #% 2022-10-04
}
\examples{
Hex2Hsl('black')
Hex2Hsl('blue')
Hex2Hsl('#FFA550')

}
\author{
SLA (based on https://bitbucket.org/mattgove/color-theory/src/master/mg_colors.py)
}
