% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{HorizontalPanelsList}
\alias{HorizontalPanelsList}
\title{Horizontal Panels List}
\usage{
HorizontalPanelsList(
  samples,
  horizontal_panels_list,
  incl_reps,
  replicate_names,
  verbosity,
  interface
)
}
\description{
Horizontal Panels List
}
\details{
Internal function: Checks/constructs a horizontal_panels_list
}
