% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{HorizontalPanelsList}
\alias{HorizontalPanelsList}
\title{Horizontal Panels List}
\usage{
HorizontalPanelsList(
  samples,
  horizontal_panels_list,
  incl_reps,
  replicate_names,
  verbosity,
  interface
)
}
\arguments{
\item{samples}{}

\item{horizontal_panels_list}{}

\item{incl_reps}{}

\item{replicate_names}{}

\item{verbosity}{}

\item{interface}{}
}
\value{

}
\description{
Internal function: 
Checks/constructs a horizontal_panels_list
}
\examples{

}
\author{
SLA
}
