% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{IsColor}
\alias{IsColor}
\title{Is Color}
\usage{
IsColor(put_color_vector)
}
\arguments{
\item{put_color_vector}{putative vector of colors}
}
\value{

}
\description{
Internal function: 
Helper function that determines whether input (vector) is a color
}
\examples{
IsColor('black')
IsColor('white')
IsColor('gnyf')
IsColor(c('white', 'blue', 'gnyf'))
IsColor('#0098AD')

}
\author{
SLA (with help from https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation/13290832#13290832)
}
\keyword{internal}
