% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{LoadAndTransformDataForTrack}
\alias{LoadAndTransformDataForTrack}
\title{Load And Transform Data For Track}
\usage{
LoadAndTransformDataForTrack(
  seqtype,
  plotted_region,
  samples,
  bigwigs,
  bigwig_dirs,
  parameters = NULL,
  get_subsamples = FALSE,
  print_order = FALSE,
  verbosity
)
}
\arguments{
\item{seqtype}{}

\item{plotted_region}{}

\item{samples}{}

\item{bigwigs}{}

\item{bigwig_dirs}{}

\item{parameters}{}

\item{get_subsamples}{}

\item{print_order}{}

\item{verbosity}{}
}
\description{
Internal function:
}
\author{
SLA
}
\keyword{internal}
