% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{LoadTracks}
\alias{LoadTracks}
\title{Load Tracks}
\usage{
LoadTracks(
  plotted_region,
  samples,
  bigwigs,
  bigwig_dirs,
  parameters,
  verbosity
)
}
\description{
Load Tracks
}
\details{
Internal function:
}
