% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OpenOptionsTable}
\alias{OpenOptionsTable}
\title{Open Options Table}
\usage{
OpenOptionsTable()
}
\value{
placeholder
}
\description{
Open the Options table (Excel sheet with info for app)
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
