% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrderedSplit}
\alias{OrderedSplit}
\title{Ordered Split}
\usage{
OrderedSplit(x, f, drop = TRUE)
}
\arguments{
\item{x}{}

\item{f}{}

\item{drop}{}
}
\value{

}
\description{
Internal function: 
Split data frame by values in a column but maintain order as order of appearance. 
Different to base R split.data.frame where order of split objects are ordered as default factors.
}
\note{
change throughout from ordered_split to OrderedSplit - done? #% 2022-10-04
}
\examples{

}
\author{
MS
}
