% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrganizeAnnotatedFeaturesInRegion}
\alias{OrganizeAnnotatedFeaturesInRegion}
\title{Organize Annotated Features In Region}
\usage{
OrganizeAnnotatedFeaturesInRegion(plotted_region, annotations)
}
\arguments{
\item{plotted_region}{}

\item{annotations}{}
}
\value{

}
\description{
Internal function: 
Wrapper function that executes the three above functions
}
\examples{

}
\author{
SLA
}
