% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrganizeOverlappingIVs}
\alias{OrganizeOverlappingIVs}
\title{Organize Overlapping IVs}
\usage{
OrganizeOverlappingIVs(annotation_gr, gap = -1L)
}
\description{
Organize Overlapping IVs
}
\details{
Internal function: Overlapping feature intervals (transcripts/genes/transcription units) are organized in separate lines to allow "non-overlapping" plotting
}
