% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrganizeOverlappingIVs}
\alias{OrganizeOverlappingIVs}
\title{Organize Overlapping Intervals}
\usage{
OrganizeOverlappingIVs(annotation_gr, gap = -1L)
}
\arguments{
\item{annotation_gr}{}

\item{gap}{}
}
\value{

}
\description{
Internal function: 
Overlapping feature intervals (transcripts/genes/transcription units) are organized in separate lines to allow "non-overlapping" plotting
}
\examples{

}
\author{
SLA
}
