% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{OrganizePanelsDimensions}
\alias{OrganizePanelsDimensions}
\title{Organize Panels Dimensions}
\usage{
OrganizePanelsDimensions(
  datasets,
  min_word_length,
  replicate_names,
  print_one_line_sample_names,
  incl_first_panel,
  plot_height_parameters,
  feature_names_font_size,
  font_size_range,
  recommended_font_sizes,
  scale_font_size,
  horizontal_panels_list,
  panel_font_size_list,
  panels_list,
  plot_widths_cm,
  panel_separators,
  strand,
  both_strands,
  strands_intermingled,
  stranded_samples,
  fixed_panel_width = FALSE,
  verbosity
)
}
\description{
Organize Panels Dimensions
}
\details{
Internal function: The overall panel dimensions can fixed upfront or it can be left open for the function to determine the dimension - with some initial parameters defined.
}
