% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PanelFontSizeList}
\alias{PanelFontSizeList}
\title{Panel Font Size List}
\usage{
PanelFontSizeList(
  samples,
  panel_font_sizes,
  panel_font_size_list,
  verbosity,
  interface
)
}
\description{
Panel Font Size List
}
\details{
Internal function: Constructs a panel_font_size_list (a font size provided to each panel)
}
