% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PanelFontSizeList}
\alias{PanelFontSizeList}
\title{Panel Font Size List}
\usage{
PanelFontSizeList(
  samples,
  panel_font_sizes,
  panel_font_size_list,
  incl_reps,
  replicate_names,
  verbosity,
  interface
)
}
\arguments{
\item{samples}{}

\item{panel_font_sizes}{}

\item{panel_font_size_list}{}

\item{incl_reps}{}

\item{replicate_names}{}

\item{verbosity}{}

\item{interface}{}
}
\value{
placeholder
}
\description{
Internal function: 
Constructs a panel_font_size_list (a font size provided to each panel)
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
