% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{plot.seqNdisplayRSession}
\alias{plot.seqNdisplayRSession}
\title{plot seq'N'display'R Session}
\usage{
\method{plot}{seqNdisplayRSession}(session, ...)
}
\arguments{
\item{session}{object of class seqNdisplayRSession}

\item{...}{arguments passed to seqNdisplay. Should contain at least the argument feature or locus.}
}
\value{
On-screen plot or pdf
}
\description{
Container for seqNdisplayR session information
}
\details{
see seqNdisplay for details. Session contains samples, colors, bigwigs, bigwig_dirs, parameters and annotation information.
}
\examples{
xl_fname = system.file('extdata', 'seqNdisplayR_sample_sheet_elaborate2.xlsx', package='seqNdisplayR')
session = LoadExcel(xl_fname, load_annotations =F)
class(session) # 'seqNdisplayRSession'
plot(session, feature='TAF1D')

}
\author{
MS (minor additions by SLA)
}
