% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayRSession.R
\name{plot.seqNdisplayRSession}
\alias{plot.seqNdisplayRSession}
\title{Plot seqNdisplayR Object}
\usage{
\method{plot}{seqNdisplayRSession}(session, ...)
}
\arguments{
\item{session}{object of class seqNdisplayR}

\item{...}{arguments passed to seqNdisplay. Should contain at least arg feature or locus.}
}
\value{
Plot or produces pdf with default settings
}
\description{
Plot seqNdisplayR Object
}
\details{
see seqNdisplay for details. Session contains samples, colors, bigwigs,
  bigwig_dirs, parameters and annotation information.
}
\examples{
xl_fname <- system.file('extdata', 'example_excel_template.xls', package='seqNdisplayR')
session <- load_excel(xl_fname, load_annotations =F)
class(session) <- 'seqNdisplayRSession'
plot(session, feature='TAF1D')

}
