% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotAnnotation}
\alias{PlotAnnotation}
\title{Plot Annotation}
\usage{
PlotAnnotation(
  annot_info,
  stranded,
  annot_cols,
  annotation_packing,
  plotted_region,
  plotted_strand,
  substrand,
  basic_plot_parameters,
  plot_start,
  plot_end,
  plot_width,
  bin_size,
  reverse_strand_direction,
  incl_feature_names,
  feature_names_above,
  incl_feature_brackets,
  incl_feature_shadings,
  feature_shading_colors,
  feature_shading_alpha,
  plot_width_parameters,
  plot_vertical_parameters,
  final_feature_text_org,
  windows_height,
  feature_font_size,
  annotation_panel_font_size,
  annot_panel_dist = 0.4,
  coords_tracks,
  font_colors,
  font_family,
  first_plot,
  scaling_factor,
  verbosity
)
}
\description{
Plot Annotation
}
\details{
Internal function:
}
