% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotData}
\alias{PlotData}
\title{Plot Data}
\usage{
PlotData(
  plotting_segment,
  plot_mat,
  colors,
  strands_alpha,
  sample_subset,
  windows_height,
  coords_tracks,
  coords_scale,
  first_plot,
  neg_vals_neg_strand,
  plotted_strand,
  y_par,
  plot_start,
  plot_end,
  bin_width,
  group_autoscale,
  incl_track_scales,
  scientific_scale,
  scale_font_size,
  full_width_cm,
  font_colors,
  font_family,
  scaling_factor,
  letter_widths,
  enhance_signals,
  scale_warning = NULL,
  verbosity
)
}
\description{
Plot Data
}
\details{
Internal function:
}
