% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotData}
\alias{PlotData}
\title{Plot Data}
\usage{
PlotData(
  plotting_segment,
  plot_mat,
  colors,
  strands_alpha,
  intermingled_color,
  sample_subset,
  windows_height,
  coords_tracks,
  coords_scale,
  first_plot,
  neg_vals_neg_strand,
  plotted_strand,
  y_par,
  plot_start,
  plot_end,
  bin_width,
  group_autoscale,
  incl_track_scales,
  scientific_scale,
  scale_font_size,
  log2transformed,
  full_width_cm,
  font_colors,
  font_family,
  scaling_factor,
  letter_widths,
  enhance_signals,
  scale_warning = NULL,
  verbosity
)
}
\arguments{
\item{plotting_segment}{}

\item{plot_mat}{}

\item{colors}{}

\item{strands_alpha}{}

\item{intermingled_color}{}

\item{sample_subset}{}

\item{windows_height}{}

\item{coords_tracks}{}

\item{coords_scale}{}

\item{first_plot}{}

\item{neg_vals_neg_strand}{}

\item{plotted_strand}{}

\item{y_par}{}

\item{plot_start}{}

\item{plot_end}{}

\item{bin_width}{}

\item{group_autoscale}{}

\item{incl_track_scales}{}

\item{scientific_scale}{}

\item{scale_font_size}{}

\item{log2transformed}{}

\item{full_width_cm}{}

\item{font_colors}{}

\item{font_family}{}

\item{scaling_factor}{}

\item{letter_widths}{}

\item{enhance_signals}{}

\item{scale_warning}{}

\item{verbosity}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
\keyword{internal}
