% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotMatrix}
\alias{PlotMatrix}
\title{Plot Matrix}
\usage{
PlotMatrix(
  plotted_region,
  basic_plot_parameters,
  plot_start,
  plot_end,
  plot_width,
  bin_size,
  actual_strand_direction,
  sample_subset,
  dummy_plot,
  tracks,
  plotting_segment,
  bin_stats
)
}
\description{
Plot Matrix
}
\details{
Internal function:
}
