% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotMatrix}
\alias{PlotMatrix}
\title{Plot Matrix}
\usage{
PlotMatrix(
  plotted_region,
  basic_plot_parameters,
  plot_start,
  plot_end,
  plot_width,
  bin_size,
  reverse_strand_direction,
  sample_subset,
  dummy_plot,
  tracks,
  plotting_segment,
  bin_stats
)
}
\arguments{
\item{plotted_region}{}

\item{basic_plot_parameters}{}

\item{plot_start}{}

\item{plot_end}{}

\item{plot_width}{}

\item{bin_size}{}

\item{reverse_strand_direction}{}

\item{sample_subset}{}

\item{dummy_plot}{}

\item{tracks}{}

\item{plotting_segment}{}

\item{bin_stats}{}
}
\description{
Internal function:
}
\author{
SLA
}
\keyword{internal}
