% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotPanels}
\alias{PlotPanels}
\title{Plot Panels}
\usage{
PlotPanels(
  plotting_segment,
  plotted_strand,
  panel_info,
  panels_list,
  panel_separators,
  separators_lwds,
  separators_colors,
  incl_first_panel,
  print_one_line_sample_names,
  replicate_names,
  plot_width_parameters,
  windows_height,
  vertical_slots,
  segment_top,
  full_width_cm,
  font_color,
  font_family,
  colors,
  first_plot,
  letter_heights,
  scaling_factor
)
}
\description{
Plot Panels
}
\details{
Internal function:
}
