% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotScale}
\alias{PlotScale}
\title{Plot Scale}
\usage{
PlotScale(
  windows_height,
  n_segment,
  coords_tracks,
  full_width_cm,
  genomic_scale_on_top,
  plot_width,
  plot_start,
  plot_end,
  first_plot,
  font_color,
  font_family,
  genomic_scale_font_size,
  scaling_factor,
  letter_widths
)
}
\description{
Plot Scale
}
\details{
Internal function:
}
