% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotScale}
\alias{PlotScale}
\title{Plot Scale}
\usage{
PlotScale(
  windows_height,
  n_segment,
  coords_tracks,
  full_width_cm,
  genomic_scale_on_top,
  plot_width,
  plot_start,
  plot_end,
  first_plot,
  font_color,
  font_family,
  genomic_scale_font_size,
  scaling_factor
)
}
\arguments{
\item{windows_height}{}

\item{n_segment}{}

\item{coords_tracks}{}

\item{full_width_cm}{}

\item{genomic_scale_on_top}{}

\item{plot_width}{}

\item{plot_start}{}

\item{plot_end}{}

\item{first_plot}{}

\item{font_color}{}

\item{font_family}{}

\item{genomic_scale_font_size}{}

\item{scaling_factor}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
