% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotSegment}
\alias{PlotSegment}
\title{Plot Segment}
\usage{
PlotSegment(
  feature,
  plotted_region,
  plotted_strand,
  both_strands,
  plotting_segment,
  basic_plot_parameters,
  neg_vals_neg_strand,
  plot_width_parameters,
  plot_vertical_parameters,
  annot_info,
  panel_info,
  panels_list,
  panel_separators,
  separators_lwds,
  separators_colors,
  incl_first_panel,
  print_one_line_sample_names,
  replicate_names,
  header,
  header_font_sizes,
  scaling_factor,
  full_width_cm,
  genomic_scale_on_top,
  genomic_scale_font_size,
  actual_strand_direction,
  bin_stats,
  dummy_plot,
  tracks,
  strands_alpha,
  unstranded_beds,
  annotation_packing,
  annotation_panel_font_size,
  incl_feature_names,
  feature_font_size,
  feature_names_above,
  final_feature_text_org,
  incl_feature_brackets,
  incl_feature_shadings,
  feature_shading_colors,
  feature_shading_alpha,
  annot_cols,
  incl_track_scales,
  scientific_scale,
  scale_font_size,
  force_scale_list,
  colors,
  alternating_background,
  bgr_colors,
  bgr_alpha,
  font_colors,
  letter_widths,
  letter_heights,
  enhance_signals,
  first_plot,
  verbosity
)
}
\description{
Plot Segment
}
\details{
Internal function:
}
