% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotSegment}
\alias{PlotSegment}
\title{Plot Segment}
\usage{
PlotSegment(
  feature,
  plotted_region,
  plotted_strand,
  both_strands,
  plotting_segment,
  basic_plot_parameters,
  neg_vals_neg_strand,
  plot_width_parameters,
  plot_vertical_parameters,
  annot_info,
  panel_info,
  panels_list,
  panel_separators,
  separators_lwds,
  separators_colors,
  incl_first_panel,
  print_one_line_sample_names,
  replicate_names,
  header,
  header_font_sizes,
  scaling_factor,
  full_width_cm,
  genomic_scale_on_top,
  genomic_scale_font_size,
  reverse_strand_direction,
  bin_stats,
  dummy_plot,
  tracks,
  strands_alpha,
  intermingled_color,
  unstranded_beds,
  annotation_packing,
  annotation_panel_font_size,
  incl_feature_names,
  feature_font_size,
  feature_names_above,
  final_feature_text_org,
  incl_feature_brackets,
  incl_feature_shadings,
  feature_shading_colors,
  feature_shading_alpha,
  annot_cols,
  group_autoscale,
  incl_track_scales,
  scientific_scale,
  scale_font_size,
  force_scale_list,
  log2transformed,
  colors,
  alternating_background,
  bgr_colors,
  bgr_alpha,
  font_colors,
  letter_widths,
  letter_heights,
  enhance_signals,
  first_plot,
  verbosity
)
}
\arguments{
\item{feature}{}

\item{plotted_region}{}

\item{plotted_strand}{}

\item{both_strands}{}

\item{plotting_segment}{}

\item{basic_plot_parameters}{}

\item{neg_vals_neg_strand}{}

\item{plot_width_parameters}{}

\item{plot_vertical_parameters}{}

\item{annot_info}{}

\item{panel_info}{}

\item{panels_list}{}

\item{panel_separators}{}

\item{separators_lwds}{}

\item{separators_colors}{}

\item{incl_first_panel}{}

\item{print_one_line_sample_names}{}

\item{replicate_names}{}

\item{header}{}

\item{header_font_sizes}{}

\item{scaling_factor}{}

\item{full_width_cm}{}

\item{genomic_scale_on_top}{}

\item{genomic_scale_font_size}{}

\item{reverse_strand_direction}{}

\item{bin_stats}{}

\item{dummy_plot}{}

\item{tracks}{}

\item{strands_alpha}{}

\item{intermingled_color}{}

\item{unstranded_beds}{}

\item{annotation_packing}{}

\item{annotation_panel_font_size}{}

\item{incl_feature_names}{}

\item{feature_font_size}{}

\item{feature_names_above}{}

\item{final_feature_text_org}{}

\item{incl_feature_brackets}{}

\item{incl_feature_shadings}{}

\item{feature_shading_colors}{}

\item{feature_shading_alpha}{}

\item{annot_cols}{}

\item{group_autoscale}{}

\item{incl_track_scales}{}

\item{scientific_scale}{}

\item{scale_font_size}{}

\item{force_scale_list}{}

\item{log2transformed}{}

\item{colors}{}

\item{alternating_background}{}

\item{bgr_colors}{}

\item{bgr_alpha}{}

\item{font_colors}{}

\item{letter_widths}{}

\item{letter_heights}{}

\item{enhance_signals}{}

\item{first_plot}{}

\item{verbosity}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
