% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlottingSegmentOrder}
\alias{PlottingSegmentOrder}
\title{Plotting Segment Order}
\usage{
PlottingSegmentOrder(
  plotting_segment_order,
  sample_names,
  header,
  include_genomic_scale,
  genomic_scale_on_top,
  incl_annot,
  horizontal_spacers
)
}
\description{
Plotting Segment Order
}
\details{
Internal function:
}
