% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlottingSegmentOrder}
\alias{PlottingSegmentOrder}
\title{Plotting Segment Order}
\usage{
PlottingSegmentOrder(
  plotting_segment_order,
  sample_names,
  header,
  include_genomic_scale,
  genomic_scale_on_top,
  incl_annot,
  horizontal_spacers
)
}
\arguments{
\item{plotting_segment_order}{}

\item{sample_names}{}

\item{header}{}

\item{include_genomic_scale}{}

\item{genomic_scale_on_top}{}

\item{incl_annot}{}

\item{horizontal_spacers}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
