% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PlotWidths}
\alias{PlotWidths}
\title{Plot Widths}
\usage{
PlotWidths(
  panels_max_width_cm,
  scale_panel_width_cm,
  margin_width_cm,
  track_width_cm,
  full_width_cm,
  incl_track_scales,
  verbosity,
  interface
)
}
\description{
Plot Widths
}
\details{
Internal function: Make sense of the supplied width arguments
}
