% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PreparePlottingInterface}
\alias{PreparePlottingInterface}
\title{Prepare Plotting Interface}
\usage{
PreparePlottingInterface(
  plot_dim,
  pdf,
  pdf_name,
  pdf_dir,
  header,
  bin_size,
  feature,
  scaling_factor
)
}
\description{
Prepare Plotting Interface
}
\details{
Internal function:
}
