% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{PreparePlottingInterface}
\alias{PreparePlottingInterface}
\title{Prepare Plotting Interface}
\usage{
PreparePlottingInterface(
  plot_dim,
  pdf,
  pdf_name,
  pdf_dir,
  header,
  bin_size,
  feature,
  scaling_factor
)
}
\arguments{
\item{plot_dim}{}

\item{pdf}{}

\item{pdf_name}{}

\item{pdf_dir}{}

\item{header}{}

\item{bin_size}{}

\item{feature}{}

\item{scaling_factor}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
