% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayRSession.R
\name{print.seqNdisplayRSession}
\alias{print.seqNdisplayRSession}
\title{Print seqNdisplayRSession object}
\usage{
\method{print}{seqNdisplayRSession}(session, verbose = FALSE, ...)
}
\arguments{
\item{session}{seqNdisplayRSession object}

\item{verbose}{print detailed information? default=FALSE}

\item{...}{arguments to glimpse_session}
}
\value{
Print to R session.
}
\description{
Prints an overview over samples, colors and associated bigwigs in a seqNdisplayR Session.
}
\details{
Convenience function for checking parsing of samples, colors and bigwigs. See glimpse_session for details.
}
\examples{
xl_fname <- system.file('extdata', 'example_excel_template.xls', package='seqNdisplayR')
session <- load_excel(xl_fname, load_annotations =F)
print(session)
print(session, verbose=T)

}
