% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{ReadInAnnotations}
\alias{ReadInAnnotations}
\title{Read In Annotations}
\usage{
ReadInAnnotations(annots, verbosity)
}
\description{
Read In Annotations
}
\details{
Internal function: Read in annotations as GRanges objects from bed files - stored in list format
}
