% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{RecommendedFontSizes}
\alias{RecommendedFontSizes}
\title{Recommended Font Sizes}
\usage{
RecommendedFontSizes(
  est_track_height_cm,
  est_min_annot_height,
  plot_vertical_parameters
)
}
\description{
Recommended Font Sizes
}
\details{
Internal function:
}
