% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{RegionGRanges}
\alias{RegionGRanges}
\title{Region GRanges}
\usage{
RegionGRanges(
  locus,
  tracks_width,
  feature = NULL,
  annotations = NULL,
  bin_start = NULL,
  extra_space = c(0.1, 0.1),
  verbosity,
  interface
)
}
\description{
Region GRanges
}
\details{
Internal function: Convert feature name or locus to be plotted to a GRanges object
}
