% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{RelativeAnnotationHeight}
\alias{RelativeAnnotationHeight}
\title{Relative Annotation Height}
\usage{
RelativeAnnotationHeight(
  annot_info,
  annot_heights,
  letter_heights,
  incl_feature_names,
  feature_text_org,
  annotation_packing,
  incl_feature_brackets,
  stranded_beds
)
}
\description{
Relative Annotation Height
}
\details{
Internal function:
}
