% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{RelativeAnnotationHeight}
\alias{RelativeAnnotationHeight}
\title{RelativeAnnotationHeight}
\usage{
RelativeAnnotationHeight(
  annot_info,
  annot_heights,
  letter_heights,
  incl_feature_names,
  feature_text_org,
  annotation_packing,
  incl_feature_brackets,
  stranded_beds
)
}
\arguments{
\item{annot_info}{}

\item{annot_heights}{}

\item{letter_heights}{}

\item{incl_feature_names}{}

\item{feature_text_org}{}

\item{annotation_packing}{}

\item{incl_feature_brackets}{}

\item{stranded_beds}{}
}
\description{
Internal function:
}
\author{
SLA
}
\keyword{internal}
