% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{run_seqNdisplayR_app}
\alias{run_seqNdisplayR_app}
\title{run seq'N'display'R app}
\usage{
run_seqNdisplayR_app(...)
}
\value{
placeholder
}
\description{
Run the shiny app
}
\examples{
run_seqNdisplayR_app()
run_seqNdisplayR_app(launch.browser = TRUE)

}
\author{
SLA
}
