% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{ScrutinizeExpandAndNameParameter}
\alias{ScrutinizeExpandAndNameParameter}
\title{Scrutinize Expand And Name Parameter}
\usage{
ScrutinizeExpandAndNameParameter(
  parameter,
  name_vector,
  use_names = FALSE,
  default_value = NULL,
  expect_standard = NULL,
  expect = NULL,
  revert_to_default = FALSE,
  alt_par_name = NULL,
  verbosity
)
}
\description{
Scrutinize Expand And Name Parameter
}
\details{
Internal function: Helper function that will expand a common "one-dimensional" parameter to fit all the names of a named object
}
