% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{SegmentTop}
\alias{SegmentTop}
\title{Segment Top}
\usage{
SegmentTop(
  plotting_segment,
  plotted_strand,
  windows_height,
  annot_info,
  dummy_plot,
  tracks,
  unstranded_beds_names,
  verbosity
)
}
\arguments{
\item{plotting_segment}{}

\item{plotted_strand}{}

\item{windows_height}{}

\item{annot_info}{}

\item{dummy_plot}{}

\item{tracks}{}

\item{unstranded_beds_names}{}

\item{verbosity}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
\keyword{internal}
