% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{SegmentTop}
\alias{SegmentTop}
\title{Segment Top}
\usage{
SegmentTop(
  plotting_segment,
  plotted_strand,
  windows_height,
  annot_info,
  dummy_plot,
  tracks,
  unstranded_beds_names,
  verbosity
)
}
\description{
Segment Top
}
\details{
Internal function:
}
