% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{seqNdisplay}
\alias{seqNdisplay}
\title{seq'N'display}
\usage{
seqNdisplay(
  datasets,
  colors,
  bigwig_dirs,
  bigwigs,
  parameters,
  plotting_segment_order = NULL,
  preloaded_tracks = NULL,
  output_tracks = FALSE,
  output_parameters = FALSE,
  input_parameters = NULL,
  both_strands = TRUE,
  strands_intermingled = TRUE,
  neg_vals_neg_strand = TRUE,
  reverse_strand_direction = FALSE,
  alternating_background = TRUE,
  bgr_colors = c("#C1B49A", "#F1F1F2"),
  bgr_alpha = 0.2,
  strands_alpha = c(100, 100),
  intermingled_color = "same",
  feature = NULL,
  locus = NULL,
  extra_space = c(1.5, 1.5),
  annots = NULL,
  annotation_packing = "collapsed2",
  annot_cols = NULL,
  annot_panel_color = "steelblue",
  annot_panel_font_size = NULL,
  bin_start = NULL,
  bin_size = "automatic",
  bins_per_cm = 250,
  track_width_cm = 10,
  full_width_cm = NULL,
  full_height_cm = NULL,
  track_height_cm = 0.3,
  title_field_height_cm = 0.66,
  genomic_scale_height_cm = 0.24,
  annotation_height_cm = 0.24,
  spacer_height_cm = 0.06,
  panels_max_width_cm = "auto",
  margin_width_cm = 0.05,
  fixed_panel_width = FALSE,
  horizontal_panels_list = NULL,
  panel_font_sizes = NULL,
  panel_font_size_list = NULL,
  panel_text_colors = c("darkgreen", "black"),
  horizontal_spacers = TRUE,
  panel_separators = c(FALSE, TRUE),
  separators_lwds = c(0.5, 1, 0.5),
  separators_colors = "black",
  incl_first_panel = TRUE,
  print_one_line_sample_names = FALSE,
  replicate_names = "rep",
  group_autoscale = TRUE,
  incl_track_scales = TRUE,
  scientific_scale = c("allow", "all", "none")[1],
  force_scale = NULL,
  scale_font_size = NULL,
  scale_panel_width_cm = "auto",
  scale_font_color = "darkred",
  header = NULL,
  suppress_header = FALSE,
  header_font_sizes = NULL,
  header_font_colors = c("black", "darkgray", "black"),
  include_genomic_scale = TRUE,
  genomic_scale_on_top = TRUE,
  genomic_scale_font_size = NULL,
  genomic_scale_font_color = "black",
  incl_feature_names = TRUE,
  feature_names_above = FALSE,
  feature_names_alternating = TRUE,
  feature_names_font_size = NULL,
  incl_feature_brackets = TRUE,
  incl_feature_shadings = TRUE,
  feature_shading_colors = c("steelblue", "hotpink"),
  feature_shading_alpha = 0.05,
  center_of_mass = FALSE,
  feature_names_font_color = "black",
  dummy_plot = FALSE,
  pdf = FALSE,
  pdf_name = NULL,
  pdf_dir = "./testplotting",
  scaling_factor = 1,
  verbosity = "normal",
  interface = "R",
  ...
)
}
\arguments{
\item{70+_arguments}{see xxx #@}
}
\value{

}
\description{
seq'N'display
}
\details{
Main function to create a plot. See xxx #@
}
\examples{

}
