% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayRSession.R
\name{seqNdisplayRSession}
\alias{seqNdisplayRSession}
\title{seqNdisplayRSession}
\usage{
seqNdisplayRSession(
  df = NULL,
  samples = NULL,
  colors = NULL,
  bigwig_dirs = NULL,
  bigwigs = NULL,
  parameters = NULL,
  annotations = NULL,
  options = NULL,
  load_annotations = F
)
}
\arguments{
\item{df}{an optional df, overrides all other options except annotations}

\item{samples}{samples object as used by seqNdisplayR}

\item{colors}{colors object as used by seqNdisplayR}

\item{bigwig_dirs}{bigwig_dirs object as used by seqNdisplayR}

\item{bigwigs}{bigwigs object as used by seqNdisplayR}

\item{parameters}{parameters object as used by seqNdisplayR}

\item{annotations}{annotations object as used by seqNdisplayR}

\item{options}{named list of other arguments used by seqNdisplay function}

\item{load_annotations}{load annotations as GRanges? default=FALSE.}
}
\value{
A named list with slots samples, colors, bigwig_dirs, bigwigs,
  parameters and annotation_files and annots. Annots is either identical to
  annotation_files or a named list of loaded GRanges.
}
\description{
Class as container for seqNdisplayR session information.
}
\details{
seqNdisplayR session object holding above slots identical to parameters
  for seqNdisplay. If df is provided parses information from columns, colors,
  bigwig_file, bigwig_directory, dataset and subgroup_1, subgroup_2 etc. See
  Excel template sheet in
  \code{system.file('extdata','example_excel_template.xls',
  package='seqNdisplayR')} for more information. The df here is tidy, ie all
  "empty" slots are filled, see \link[seqNdisplayR]{fill_df}. Otherwise see
  \code{vignette(package='seqNdisplayR')}. If options are not provided, adds
  default options to session object. If load_annotations=TRUE will try to
  load annotations using \link[rtracklayer]{import}. UPS: Currently only bed
  files are used correctly by
  seqNdisplayR.
}
