% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{Session2Df}
\alias{Session2Df}
\title{Session 2 Df}
\usage{
Session2Df(
  .samples,
  .colors,
  .bigwigs,
  .bigwig_dirs,
  .parameters,
  strand_regex = c(`+` = "plus", `-` = "minus"),
  factorize = FALSE,
  level = 0
)
}
\arguments{
\item{.samples}{samples object as used by seqNdisplayR}

\item{.colors}{colors object as used by seqNdisplayR}

\item{.bigwigs}{bigwigs object as used by seqNdisplayR}

\item{.bigwig_dirs}{bigwig_dirs as used in seqNdisplayR}

\item{.parameters}{list of parameters as used in seqNdisplayR}

\item{strand_regex}{named vector c('+': ..., '-': ...) for regex for converting plus strand to minus strand bigwig names}

\item{factorize}{TRUE/FALSE. Default: FALSE}

\item{level}{do not change, required internally during recursion, defaults=0.}
}
\value{
A dataframe with columns: color, bigwig_directory, bigwig_file, strand, batch, dataset and optionally subgroup_1, subgroup_2, ...
}
\description{
Internal function: 
Converts session information to data frame as in Excel import sheet.
}
\details{
Converts session information to dataframe as specified in Excel
  import sheet but using all-filled mode. Batch information is obtained from parameters.
}
\examples{

}
\author{
MS
}
\keyword{internal}
