% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{Session2xlsx}
\alias{Session2xlsx}
\title{Session2xlsx}
\usage{
Session2xlsx(session, path, ...)
}
\arguments{
\item{session}{seqNdisplayRSession object}

\item{path}{Excel file path}

\item{...}{}
}
\value{
Write Excel template as described in \link{LoadExcel}.
}
\description{
Save session object to Excel xlsx-file
}
\note{
change throughout from session_to_xlsx to Session2xlsx - done? #% 2022-10-04
}
\examples{

}
\author{
MS (minor additions by SLA)
}
