% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{SortUnlistedSampleNames}
\alias{SortUnlistedSampleNames}
\title{Sort Unlisted Sample Names}
\usage{
SortUnlistedSampleNames(
  unlisted_which_sample,
  unlisted_sample_names,
  incl_rep = F
)
}
\arguments{
\item{unlisted_which_sample}{}

\item{unlisted_sample_names}{}

\item{incl_rep}{}
}
\value{

}
\description{
Internal function: 
Sort Unlisted Sample Names
}
\examples{

}
\author{
SLA
}
\keyword{internal}
