% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{UnpackSamples}
\alias{UnpackSamples}
\title{Unpack Samples}
\usage{
UnpackSamples(
  seqtype,
  samples,
  which_samples,
  which_reps,
  bigwig_list,
  bigwig_dirs,
  verbosity
)
}
\description{
Unpack Samples
}
\details{
Internal function:
}
