% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{UnpackSamples}
\alias{UnpackSamples}
\title{Unpack Samples}
\usage{
UnpackSamples(
  seqtype,
  samples,
  which_samples,
  which_reps,
  bigwig_list,
  bigwig_dirs,
  verbosity
)
}
\arguments{
\item{seqtype}{}

\item{samples}{}

\item{which_samples}{}

\item{which_reps}{}

\item{bigwig_list}{}

\item{bigwig_dirs}{}

\item{verbosity}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
