% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{UpdatePlotVerticalParameters}
\alias{UpdatePlotVerticalParameters}
\title{Update Plot Vertical Parameters}
\usage{
UpdatePlotVerticalParameters(
  plot_vertical_parameters,
  track_height_cm_estimate,
  title_field_height_cm,
  genomic_scale_height_cm,
  annotation_height_cm,
  spacer_height_cm
)
}
\description{
Update Plot Vertical Parameters
}
\details{
Internal function:
}
