% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{UpdateTrackVector}
\alias{UpdateTrackVector}
\title{Update Track Vector}
\usage{
UpdateTrackVector(track_vector, plot_vertical_parameters)
}
\description{
Update Track Vector
}
\details{
Internal function:
}
