% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{UpdateTrackVector}
\alias{UpdateTrackVector}
\title{Update Track Vector}
\usage{
UpdateTrackVector(track_vector, plot_vertical_parameters)
}
\arguments{
\item{track_vector}{}

\item{plot_vertical_parameters}{}
}
\value{
placeholder
}
\description{
Internal function:
}
\examples{
NULL

}
\author{
SLA
}
\keyword{internal}
