% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqNdisplayR.R
\name{YParameters}
\alias{YParameters}
\title{Y Parameters}
\usage{
YParameters(plot_mat, plotting_segment, force_scale_list, group_autoscale)
}
\arguments{
\item{plot_mat}{}

\item{plotting_segment}{}

\item{force_scale_list}{}

\item{group_autoscale}{}
}
\value{

}
\description{
Internal function:
}
\examples{

}
\author{
SLA
}
